+++ test-studio/sanity.config.js
@@ -1,6 +1,7 @@
 import {visionTool} from '@sanity/vision'
 import {defineConfig} from 'sanity'
 import {digitalOceanFiles} from 'sanity-plugin-digital-ocean-files'
+import {s3Files} from 'sanity-plugin-s3-files'
 import {deskTool} from 'sanity/desk'
 import {schemaTypes} from './schemas'
 
@@ -15,20 +16,33 @@ export default defineConfig({
     deskTool(),
     visionTool(),
     digitalOceanFiles({
-      toolTitle: 'Media Library',
+      toolTitle: 'Digital Ocean',
       // If you want to restrict file types library-wide:
       // defaultAccept: {
       //   'application/pdf': ['pdf'],
       //   'video/*': ['mp4', 'mov', 'webm'],
       // },
       credentials: {
-        bucketKey: process.env.SANITY_STUDIO_BUCKET_KEY,
-        bucketRegion: process.env.SANITY_STUDIO_BUCKET_REGION,
-        getSignedUrlEndpoint: process.env.SANITY_STUDIO_SIGNED_URL_ENDPOINT,
-        deleteObjectEndpoint: process.env.SANITY_STUDIO_DELETE_OBJECT_ENDPOINT,
-        folder: process.env.SANITY_STUDIO_UPLOAD_FOLDER,
+        bucketKey: process.env.SANITY_STUDIO_DO_BUCKET_KEY,
+        bucketRegion: process.env.SANITY_STUDIO_DO_BUCKET_REGION,
+        getSignedUrlEndpoint: process.env.SANITY_STUDIO_DO_SIGNED_URL_ENDPOINT,
+        deleteObjectEndpoint: process.env.SANITY_STUDIO_DO_DELETE_OBJECT_ENDPOINT,
+        folder: process.env.SANITY_STUDIO_DO_UPLOAD_FOLDER,
         subdomain: null,
         secretForValidating: null,
+      },
+    }),
+    s3Files({
+      toolTitle: 'S3',
+      // If you want to restrict file types library-wide:
+      // defaultAccept: {
+      //   '*': ['*'],
+      // },
+      credentials: {
+        bucketKey: process.env.SANITY_STUDIO_S3_BUCKET_KEY,
+        bucketRegion: process.env.SANITY_STUDIO_S3_BUCKET_REGION,
+        getSignedUrlEndpoint: process.env.SANITY_STUDIO_S3_SIGNED_URL_ENDPOINT,
+        deleteObjectEndpoint: process.env.SANITY_STUDIO_S3_DELETE_OBJECT_ENDPOINT,
       },
     }),
   ],