+++ test-studio/CHANGELOG.md
@@ -1,5 +1,14 @@
 # test-studio
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - sanity-plugin-digital-ocean-files@1.0.1
+  - sanity-plugin-external-files@1.0.1
+  - sanity-plugin-s3-files@1.0.1
+
 ## 1.0.1
 
 ### Patch Changes